+++ server/src/services/mod.rs
@@ -1,5 +1,2 @@
-pub mod push;
-pub mod onesignal;
-pub mod fcm;
 pub mod ntfy;
 pub mod apns;